Implement large chunks of Intellij handling.
Unlike previous Workspace tool versions where we had 4 project models:
- One for the Gradle data exported from Gradle.
- One for the manipulatable model inside the main wstool app, plus its interfaces.
- One for the 'workspace' representation, mostly controlled by the workspace module.
- One within the workspace module prior to emitting it.

We instead, have exactly 2 project models.
- One for the gradle extraction data.
- One abstraction interface, that the main app can use, implemented by the workspace modules.

The intellij module implements these modules as it requires for emitting the project. That being, source sets are actually modules in Intellij's implementation of the api, however, this is not exposed back to the main application.